Rename project to import-analyzer-py
Renamed from remove-unused-imports to import-analyzer-py to better
reflect the expanded scope of the tool:
- Cross-file import analysis
- Re-export tracking
- Cascade detection
- Circular import warnings
- Unreachable file detection

Changes:
- Renamed module directory: remove_unused_imports/ -> import_analyzer/
- Updated package name in pyproject.toml
- Updated CLI entry point: remove-unused-imports -> import-analyzer
- Updated all internal imports
- Updated all test imports
- Updated README.md and CLAUDE.md
