+++ import_analyzer/__main__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from remove_unused_imports._main import main
+from import_analyzer._main import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
